Extract settings helpers and sync UI
